Try with abstract class FileStorage
