+++ src/computer/hardware/drive/Drive.java
@@ -4,6 +4,7 @@ import computer.software.file.File;
 
 public interface Drive  extends Components {
     void addFile(File file);
+    void removeFile(File file);
     void listFiles();
     File findFile(String fileName);
 }
