+++ src/computer/software/file/FileStorage.java
@@ -1,41 +1,37 @@
-package computer.hardware.usbdevice;
+package computer.software.file;
 
 import computer.hardware.Capacity;
 import computer.hardware.drive.Drive;
-import computer.software.file.File;
-import computer.software.file.FileNotFoundException;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class MemoryStick extends USBDevice {
+public abstract class FileStorage implements Drive {
 
-    private final Capacity STORAGE_CAPACITY;
-    private Integer capacityUsage = 0;
+    protected final Capacity STORAGE_CAPACITY;
+    protected Integer capacityUsage = 0;
+    protected List<File> files = new ArrayList<>();
 
-    private List<File> files = new ArrayList<>();
+    protected FileStorage(Capacity storageCapacity) {
+        STORAGE_CAPACITY = storageCapacity;
+    }
 
-    public MemoryStick(String name, Capacity storageCapacity) {
-        super(name);
-        this.STORAGE_CAPACITY = storageCapacity;
+    protected Capacity getStorageCapacity() {
+        return STORAGE_CAPACITY;
     }
 
     public void addFile(File file) {
-        if (connected) {
             if(STORAGE_CAPACITY.getSize() > capacityUsage + file.getSize()) {
                 files.add(file);
                 capacityUsage+=file.getSize();
             } else {
                 System.out.println("Not enough space in storage capacity");     //TODO utworzyć wyjątek i wykorzystać tutaj
             }
-        } else {
-            throw new MemoryStickDisconnectedException("Memory Stick is disconnected");
-        }
     }
 
     public void removeFile(File file) {
         try{
-            findFileByName(file.getName());
+            findFile(file.getName());
             files.remove(file);
             capacityUsage-=file.getSize();
         } catch(FileNotFoundException e) {
@@ -43,20 +39,18 @@ public class MemoryStick extends USBDevice {
         }
     }
 
-    public List<File> getFiles() {
-        return files;
+    public void listFiles() {
+        for (File file : files){
+            System.out.println(file.getName());
+        }
     }
 
-    public File findFileByName(String name) {
+    public File findFile(String name) {
         for (File file : files) {
             if (file.getName().equals(name)) {
                 return file;
             }
         }
         throw new FileNotFoundException("File " + name + " not found");
-    }
-
-    public Capacity getStorageCapacity() {
-        return STORAGE_CAPACITY;
     }
 }
